requirements.txt
Python=3.10.12
Flask==3.0.1
python-dotenv==1.0.0
python-steam-api=1.2.2
beautifulsoup4=4.12.3
